update/project_deps

+++ common/src/main/java/jp/co/soramitsu/common/data/network/OptionsProvider.kt
@@ -1,7 +1,7 @@
 package jp.co.soramitsu.common.data.network
 
 object OptionsProvider {
-    var CURRENT_VERSION_CODE: Int = 0
+    var CURRENT_VERSION_CODE: Long = 0
     var CURRENT_VERSION_NAME: String = ""
     var APPLICATION_ID: String = ""
     var CURRENT_BUILD_TYPE: String = ""
